# swagger.json
{"openapi":"3.1.0","info":{"title":"Advertisement app","version":"0.1.0"},"paths":{"/auth/jwt/login":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Login","operationId":"auth_jwt_login_auth_jwt_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_auth_jwt_login_auth_jwt_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"LOGIN_BAD_CREDENTIALS":{"summary":"Bad credentials or the user is inactive.","value":{"detail":"LOGIN_BAD_CREDENTIALS"}},"LOGIN_USER_NOT_VERIFIED":{"summary":"The user is not verified.","value":{"detail":"LOGIN_USER_NOT_VERIFIED"}}}}}},"204":{"description":"No Content"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/jwt/logout":{"post":{"tags":["auth"],"summary":"Auth:Jwt.Logout","operationId":"auth_jwt_logout_auth_jwt_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Missing token or inactive user."},"204":{"description":"No Content"}},"security":[{"APIKeyCookie":[]}]}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register:Register","operationId":"register_register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"REGISTER_USER_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"REGISTER_USER_ALREADY_EXISTS"}},"REGISTER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"REGISTER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categories/":{"get":{"tags":["categories"],"summary":"Get Categories","operationId":"get_categories_categories__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CategoryRead"},"type":"array","title":"Response Get Categories Categories  Get"}}}}}},"post":{"tags":["categories"],"summary":"Create Category","operationId":"create_category_categories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/groups/":{"get":{"tags":["groups"],"summary":"Get Groups","operationId":"get_groups_groups__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GroupRead"},"type":"array","title":"Response Get Groups Groups  Get"}}}}}},"post":{"tags":["groups"],"summary":"Create Group","operationId":"create_group_groups__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/groups/{id}/":{"get":{"tags":["groups"],"summary":"Get Group","operationId":"get_group_groups__id___get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}/":{"delete":{"tags":["groups"],"summary":"Delete Group","operationId":"delete_group_groups__group_id___delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"group_id","in":"path","required":true,"schema":{"type":"integer","title":"Group Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/":{"get":{"tags":["ads"],"summary":"Get Ads","operationId":"get_ads_ads__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Page Size"}},{"name":"category_id","in":"query","required":false,"schema":{"type":"integer","title":"Category Id"}},{"name":"type","in":"query","required":false,"schema":{"type":"string","title":"Type"}},{"name":"sort_by_category","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Sort By Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["ads"],"summary":"Create Advertisement","operationId":"create_advertisement_ads__post","security":[{"APIKeyCookie":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertisementCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/{id}/":{"get":{"tags":["ads"],"summary":"Get Ad","operationId":"get_ad_ads__id___get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["ads"],"summary":"Update Ad","operationId":"update_ad_ads__id___patch","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdvertisementCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["ads"],"summary":"Delete Ad","operationId":"delete_ad_ads__id___delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/{ad_id}/":{"get":{"tags":["complaints"],"summary":"Get Complaints","operationId":"get_complaints_ads__ad_id___get","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ComplaintRead"},"title":"Response Get Complaints Ads  Ad Id   Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["complaints"],"summary":"Create Complaint","operationId":"create_complaint_ads__ad_id___post","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ads/{ad_id}/{id}/":{"delete":{"tags":["complaints"],"summary":"Delete Complaints","operationId":"delete_complaints_ads__ad_id___id___delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"ad_id","in":"path","required":true,"schema":{"type":"integer","title":"Ad Id"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdvertisementCreate":{"properties":{"title":{"type":"string","title":"Title"},"type":{"$ref":"#/components/schemas/AdvertisementType"},"description":{"type":"string","title":"Description"},"price":{"type":"integer","title":"Price"},"group_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Group Id"},"photos":{"items":{"$ref":"#/components/schemas/PhotoCreate"},"type":"array","title":"Photos"},"category_id":{"type":"integer","title":"Category Id"}},"type":"object","required":["title","type","description","price","group_id","photos","category_id"],"title":"AdvertisementCreate"},"AdvertisementType":{"type":"string","enum":["sell","buy","service"],"title":"AdvertisementType"},"Body_auth_jwt_login_auth_jwt_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_auth_jwt_login_auth_jwt_login_post"},"CategoryCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"CategoryCreate"},"CategoryRead":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"CategoryRead"},"ComplaintCreate":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"ComplaintCreate"},"ComplaintRead":{"properties":{"id":{"type":"integer","title":"Id"},"author_id":{"$ref":"#/components/schemas/UserRead"},"advertisement_id":{"$ref":"#/components/schemas/AdvertisementCreate"},"text":{"type":"string","title":"Text"}},"type":"object","required":["id","author_id","advertisement_id","text"],"title":"ComplaintRead"},"ErrorModel":{"properties":{"detail":{"anyOf":[{"type":"string"},{"additionalProperties":{"type":"string"},"type":"object"}],"title":"Detail"}},"type":"object","required":["detail"],"title":"ErrorModel"},"GroupCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"avatar":{"type":"string","title":"Avatar"}},"type":"object","required":["title","description","avatar"],"title":"GroupCreate"},"GroupRead":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"avatar":{"type":"string","title":"Avatar"}},"type":"object","required":["id","title","description","avatar"],"title":"GroupRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PhotoCreate":{"properties":{"url":{"type":"string","title":"Url"}},"type":"object","required":["url"],"title":"PhotoCreate"},"RecallCreate":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"RecallCreate"},"RecallRead":{"properties":{"id":{"type":"integer","title":"Id"},"author_id":{"$ref":"#/components/schemas/UserRead"},"advertisement_id":{"$ref":"#/components/schemas/AdvertisementCreate"},"text":{"type":"string","title":"Text"}},"type":"object","required":["id","author_id","advertisement_id","text"],"title":"RecallRead"},"UserCreate":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true},"is_superuser":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Superuser","default":false},"is_verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Verified","default":false},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"contact":{"type":"string","title":"Contact"}},"type":"object","required":["email","password","first_name","last_name","contact"],"title":"UserCreate"},"UserRead":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"is_active":{"type":"boolean","title":"Is Active","default":true},"is_superuser":{"type":"boolean","title":"Is Superuser","default":false},"is_verified":{"type":"boolean","title":"Is Verified","default":false},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"contact":{"type":"string","title":"Contact"}},"type":"object","required":["id","email","first_name","last_name","contact"],"title":"UserRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyCookie":{"type":"apiKey","in":"cookie","name":"advertisements"}}}}